// src/app/model/common/info-scuola.ts
// Classe che mappa gli attributi di una macro funzionalità applicativa (di primo livello)
export class InfoScuola {
  codScuola: string;
  codOrdineScuola: string;
  descrizione: string;
  indirizzo: string;
  codCategoriaScuola: string;
  codCircoscrizione: string;
  annoScolastico: string;
  tempoBreve: boolean;
  tempoLungo: boolean;
  tempoIntermedio: boolean;
}